Renamed print module to display and moved to own file
